Fixes startup flashes.
Before this, the entire screen moved on every load, and the svg
was showing up as a big blue band on the screen for a split-second.
We now have some basic styling in the html to proved as much jarring
and the svg has dimensions already so it's not as huge on startup.